builder merged into cfgdata
